split `index.js` to separate files
