+++ web_src/js/features/comp/ColorPicker.js
@@ -1,0 +1,13 @@
+import createColorPicker from '../colorpicker.js';
+
+function initCompColorPicker() {
+  createColorPicker($('.color-picker'));
+
+  $('.precolors .color').on('click', function () {
+    const color_hex = $(this).data('color-hex');
+    $('.color-picker').val(color_hex);
+    $('.minicolors-swatch-color').css('background-color', color_hex);
+  });
+}
+
+export {initCompColorPicker};
