+++ web_src/js/features/repo-commit.js
@@ -1,0 +1,9 @@
+function initRepoCommitButton() {
+  $('.commit-button').on('click', function (e) {
+    e.preventDefault();
+    $(this).parent().find('.commit-body').toggle();
+  });
+}
+
+export {initRepoCommitButton};
+
